Fixed session behaviour
