+++ src/pages/_document.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { Html, Head, Main, NextScript } from 'next/document'
  
 export default function Document() {
