+++ src/pages/dashboard.tsx
@@ -1,22 +1,29 @@
 "use client";
 
-import { useAuth } from "../context/authContext"
+import { useAuth } from "../context/authContext";
 import { useLoginStore } from "@/store/store";
 import { useRouter } from "next/router";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 
 export default function Dashboard() {
-  const { loggedIn } = useLoginStore();
+  const { user, checkSession } = useAuth();
+  const [loading, setLoading] = useState(true); 
   const router = useRouter();
-  const { user } = useAuth();
-  console.log(user);
-  if (!loggedIn) {
-    useEffect(() => {
-      router.push("/initialview");
-    }, [router]);
-    return null;
-  } else {
-    return <p>Hello {user.email}</p>;
-  }
+  useEffect(() => {
+    const pullSession = async () => {
+      await checkSession();
+      setLoading(false);
+    };
+    pullSession();
+  }, []);
 
+  useEffect(() => {
+    if (!loading && user == null) {
+      router.replace("/");
+    }
+  }, [user, loading]);
+
+  if(loading) return <p></p>;
+
+  return <p>Hello {user.email}!</p>;
 }
