+++ src/pages/index.tsx
@@ -1,4 +1,50 @@
-export default function Home() {
-    return <div>Hello world!</div>
-}
+"use client";
 
+import { useFormStore } from "@/store/store";
+import { Signup } from "@/components/signup";
+import { Login } from "@/components/login";
+import { useAuth } from "@/context/authContext";
+import { useEffect, useState } from "react";
+import { useRouter } from "next/router";
+
+export default function InitialView() {
+  const { user, checkSession } = useAuth();
+  const [loading, setLoading] = useState(true); 
+  const router = useRouter();
+  useEffect(() => {
+    const pullSession = async () => {
+      await checkSession();
+      setLoading(false);
+    };
+    pullSession();
+  }, []);
+
+  useEffect(() => {
+    if (!loading && user != null) {
+      router.push("/dashboard");
+    }
+  }, [user, loading]);
+
+  const { type } = useFormStore();
+  return (
+    <>
+      <div className="flex h-screen w-screen">
+        <div className="flex h-screen flex-col w-1/2 h-screen">
+          <div className="flex flex-col m-auto mt-20 w-100 h-100 items-center ">
+            <img src="logo.png" className="w-50 h-50 mb-3" alt="" />
+            <div>
+              <h3 className="text-5xl font-bold">Get Started</h3>
+              <span className="text-gray-800/80 font-medium">
+                Log-in or create account
+              </span>
+            </div>
+            {type == "login" ? <Login /> : <Signup />}
+          </div>
+        </div>
+        <div className="flex flex-col w-1/2 h-screen relative shadow-lg shadow-gray-500">
+          <img src="/login.png" className="h-screen" alt="walpapper" />
+        </div>
+      </div>
+    </>
+  );
+}
